SERVER-28833 32-bit FTDC fixes

+++ src/mongo/util/procparser.cpp
@@ -274,7 +274,7 @@ Status parseProcStat(const std::vector<StringData>& keys,
                     value = 0;
                 }
 
-                builder->appendNumber(key, value);
+                builder->appendNumber(key, static_cast<long long>(value));
             }
         }
     }
@@ -378,9 +378,9 @@ Status parseProcMemInfo(const std::vector<StringData>& keys,
                     keyWithSuffix.append("_kb");
                 }
 
-                builder->appendNumber(keyWithSuffix, value);
+                builder->appendNumber(keyWithSuffix, static_cast<long long>(value));
             } else {
-                builder->appendNumber(key, value);
+                builder->appendNumber(key, static_cast<long long>(value));
             }
         }
     }
@@ -511,7 +511,7 @@ Status parseProcDiskStats(const std::vector<StringData>& disks,
                 BSONObjBuilder sub(builder->subobjStart(disk));
 
                 for (size_t index = 0; index < stats.size() && index < kDiskFieldCount; ++index) {
-                    sub.appendNumber(kDiskFields[index], stats[index]);
+                    sub.appendNumber(kDiskFields[index], static_cast<long long>(stats[index]));
                 }
 
                 sub.doneFast();
